+++ app/src/main/java/org/astraea/app/web/BalancerHandler.java
@@ -226,7 +226,7 @@ class BalancerHandler implements Handler, AutoCloseable {
                         solution ->
                             new PlanReport(
                                 changes.apply(solution),
-                                HasMoveCost.migrationCosts(contextCluster, solution.proposal())))
+                                MigrationCost.migrationCosts(contextCluster, solution.proposal())))
                     .orElse(null);
     var phase = balancerConsole.taskPhase(taskId).orElseThrow();
     return new PlanExecutionProgress(
